+++ CHANGELOG.md
@@ -6,7 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.4] - 2019-25-11
+## [0.3.0] - 2019-12-14
+### Added
+- Added `LicenseId::is_gnu` to indicate the GNU licenses (GPL, AGPL, LGPL, GFDL), which use a different suffix format than all other licenses
+- `std::error::Error` is now (properly) implemented for `error::ParseError`
+- `LicenseReq` not implements `From<LicenseId>`
+
+### Changed
+- `Lexer` and `Token` can now be reached via the `lexer` module
+- `parser` and `expression` are no longer part of the public API
+- `IS_COPYLEFT`, `IS_DEPRECATED`, `IS_FSF_LIBRE`, and `IS_OSI_APPROVED` are no longer part of the public API
+- The `GFDL*` licenses are now marked as copyleft
+
+### Fixed
+- When creating a `LicenseReq` from a GNU license, the license identifier is converted into its base form,
+eg. `GPL-2.0-or-later` becomes `GPL-2.0+` so that the GNU style license identifiers can be used just the same
+as all of the other ones. See [this issue](https://github.com/EmbarkStudios/cargo-deny/issues/55)
+
+## [0.2.4] - 2019-11-25
 ### Added
 - Impl Display for Licensee
 
@@ -38,7 +55,8 @@ a user provided callback
 ### Added
 - Initial add of spdx crate, based primarly on [`license-exprs`](https://github.com/rust-lang-nursery/license-exprs)
 
-[Unreleased]: https://github.com/EmbarkStudios/spdx/compare/0.2.4...HEAD
+[Unreleased]: https://github.com/EmbarkStudios/spdx/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/EmbarkStudios/spdx/compare/0.2.4...0.3.0
 [0.2.4]: https://github.com/EmbarkStudios/spdx/compare/0.2.3...0.2.4
 [0.2.3]: https://github.com/EmbarkStudios/spdx/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/EmbarkStudios/spdx/compare/0.2.1...0.2.2